requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
arrow
user-agents
psutil
python-dotenv
psycopg2
black
ruff
pre-commit
playwright
authlib
httpx
itsdangerous
alembic